+++ src/OmniSharp.DotNet/Cache/ProjectStatesCache.cs
@@ -11,51 +11,6 @@ using OmniSharp.Services;
 
 namespace OmniSharp.DotNet.Cache
 {
-    public class ProjectEntry
-    {
-        private readonly Dictionary<NuGetFramework, ProjectState> _states
-                   = new Dictionary<NuGetFramework, ProjectState>();
-
-        public ProjectEntry(string projectDirectory)
-        {
-            ProjectDirectory = projectDirectory;
-        }
-
-        public string ProjectDirectory { get; }
-
-        public IEnumerable<NuGetFramework> Frameworks => _states.Keys;
-
-        public IEnumerable<ProjectState> ProjectStates => _states.Values;
-
-        public ProjectState Get(NuGetFramework framework)
-        {
-            ProjectState result;
-            if (_states.TryGetValue(framework, out result))
-            {
-                return result;
-            }
-            else
-            {
-                return null;
-            }
-        }
-
-        public void Set(ProjectState state)
-        {
-            _states[state.ProjectContext.TargetFramework] = state;
-        }
-
-        public bool Remove(NuGetFramework framework)
-        {
-            return _states.Remove(framework);
-        }
-
-        public override string ToString()
-        {
-            return $"ProjectEntry {ProjectDirectory}, {_states.Count} states";
-        }
-    }
-
     public class ProjectStatesCache
     {
         private readonly Dictionary<string, ProjectEntry> _projects
@@ -209,14 +164,13 @@ namespace OmniSharp.DotNet.Cache
             }
         }
 
-        private void EmitProject(string eventType, DotNetProjectInformation information) {
+        private void EmitProject(string eventType, DotNetProjectInformation information)
+        {
             _emitter.Emit(
                 eventType,
                 new ProjectInformationResponse()
                 {
-                    { "DotNetProject", information },
-                    // the key is hard coded in VSCode
-                    { "DnxProject", information }
+                    { "DotNetProject", information }
                 });
         }
     }
